pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clip4cad"
version = "0.1.0"
description = "CLIP4CAD: Unified Multimodal Encoder for CAD Models"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "MMCAD Team"}
]
keywords = ["CAD", "multimodal", "CLIP", "3D", "B-Rep", "point cloud"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "einops>=0.7.0",
    "tqdm>=4.66.0",
    "wandb>=0.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
]
eval = [
    "scikit-learn>=1.3.0",
    "faiss-cpu>=1.7.0",
]
full = [
    "trimesh>=4.0.0",
    "open3d>=0.17.0",
    "tensorboard>=2.15.0",
]

[project.urls]
Homepage = "https://github.com/exanos/MMCAD"

[tool.setuptools.packages.find]
include = ["clip4cad*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]
